requirements.txt
# Streamlit and Databricks App Requirements

# Core Framework
streamlit>=1.28.0

# Databricks SQL Connector
databricks-sql-connector>=3.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Environment Management
python-dotenv>=1.0.0

# Data Visualization (optional, but useful)
plotly>=5.17.0
altair>=5.0.0

# Excel export support (optional)
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0

